+++ src/types/OrderType.ts
@@ -1,0 +1,14 @@
+export interface OrderListType {
+  id: number;
+  totalItemsPrice: number;
+  discountPrice: number;
+  orderItems: OrderItemType[];
+}
+
+export interface OrderItemType {
+  orderItemId: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+  quantity: number;
+}
