+++ src/views/Payment/components/OrderItemList/index.ts
@@ -1,0 +1,1 @@
+export { default as OrderItemList } from './OrderItemList';
