+++ src/views/Payment/components/OrderList/index.ts
@@ -1,0 +1,1 @@
+export { default as OrderList } from './OrderList';
